change signature to return reader of extracted fs. move to mutate package
