+++ v1/mutate/BUILD.bazel
@@ -20,6 +20,8 @@ go_test(
         "mutate_test.go",
     ],
     data = glob(["testdata/**"]) + [
+        ":whiteout_image.tar",
+        ":overwritten_file.tar",
         ":source_image.tar",  # keep
     ],
     embed = [":go_default_library"],
@@ -33,6 +35,7 @@ go_test(
 load(
     "@io_bazel_rules_docker//container:container.bzl",
     "container_image",
+    "container_layer",
 )
 
 container_image(
@@ -40,5 +43,49 @@ container_image(
     files = [
         "testdata/bar",
         "testdata/foo",
+    ],
+)
+
+container_layer(
+    name = "base_layer",
+    files = [
+        "testdata/whiteout/bar.txt",
+        "testdata/whiteout/foo.txt",
+    ],
+)
+
+container_layer(
+    name = "whiteout_layer",
+    files = [
+        "testdata/whiteout/.wh.foo.txt",
+    ],
+)
+
+container_image(
+    name = "whiteout_image",
+    layers = [
+        ":base_layer",
+        ":whiteout_layer",
+    ],
+)
+
+container_layer(
+    name = "unmodified_foo",
+    files = [
+        "testdata/whiteout/foo.txt",
+    ],
+)
+
+container_layer(
+    name = "symlinked_foo",
+    files = ["testdata/whiteout/bar.txt"],
+    symlinks = {"foo.txt": "bar.txt"},
+)
+
+container_image(
+    name = "overwritten_file",
+    layers = [
+        ":unmodified_foo",
+        ":symlinked_foo",
     ],
 )
