+++ v1/mutate/mutate_test.go
@@ -15,9 +15,13 @@
 package mutate
 
 import (
+	"archive/tar"
+	"bytes"
 	"fmt"
 	"io"
 	"io/ioutil"
+	"os"
+	"path/filepath"
 	"reflect"
 	"strings"
 	"testing"
@@ -26,6 +30,99 @@ import (
 	"github.com/google/go-containerregistry/v1"
 	"github.com/google/go-containerregistry/v1/tarball"
 )
+
+func TestExtractWhiteout(t *testing.T) {
+	img, err := tarball.ImageFromPath("whiteout_image.tar", nil)
+	if err != nil {
+		t.Errorf("Error loading image: %v", err)
+	}
+	tarPath, _ := filepath.Abs("img.tar")
+	defer os.Remove(tarPath)
+	flattened, err := Extract(img)
+	if err != nil {
+		t.Errorf("Error when flattening image: %v", err)
+	}
+	f, err := os.Create(tarPath)
+	if err != nil {
+		t.Errorf("Error when opening tar file: %v", err)
+	}
+	if _, err := io.Copy(f, flattened); err != nil {
+		t.Errorf("Error when writing flattened fs to disk: %v", err)
+	}
+	tr := tar.NewReader(f)
+	for {
+		header, err := tr.Next()
+		if err == io.EOF {
+			break
+		}
+		name := header.Name
+		for _, part := range filepath.SplitList(name) {
+			if part == "foo" {
+				t.Errorf("whiteout file found in tar: %v", name)
+			}
+		}
+	}
+}
+
+func TestExtractOverwrittenFile(t *testing.T) {
+	img, err := tarball.ImageFromPath("overwritten_file.tar", nil)
+	if err != nil {
+		t.Fatalf("Error loading image: %v", err)
+	}
+	tarPath, _ := filepath.Abs("img.tar")
+	defer os.Remove(tarPath)
+	f, err := os.Create(tarPath)
+	if err != nil {
+		t.Errorf("Error when opening tar file: %v", err)
+	}
+	flattened, err := Extract(img)
+	if err != nil {
+		t.Errorf("Error when flattening image: %v", err)
+	}
+	if _, err := io.Copy(f, flattened); err != nil {
+		t.Errorf("Error when writing flattened fs to disk: %v", err)
+	}
+
+	tr := tar.NewReader(f)
+	for {
+		header, err := tr.Next()
+		if err == io.EOF {
+			break
+		}
+		name := header.Name
+		if strings.Contains(name, "foo.txt") {
+			buf := new(bytes.Buffer)
+			buf.ReadFrom(tr)
+			if strings.Contains(buf.String(), "foo") {
+				t.Errorf("Contents of file were not correctly overwritten")
+			}
+		}
+	}
+}
+
+func TestWhiteoutDir(t *testing.T) {
+	fsMap := map[string]bool{
+		"baz":      true,
+		"red/blue": true,
+	}
+	var tests = []struct {
+		path     string
+		whiteout bool
+	}{
+		{"usr/bin", false},
+		{"baz/foo.txt", true},
+		{"baz/bar/foo.txt", true},
+		{"red/green", false},
+		{"red/yellow.txt", false},
+	}
+
+	for _, tt := range tests {
+		whiteout := InWhiteoutDir(fsMap, tt.path)
+		if whiteout != tt.whiteout {
+			t.Errorf("Whiteout %s: expected %v, but got %v", tt.path, tt.whiteout, whiteout)
+		}
+	}
+}
 
 func TestNoopCondition(t *testing.T) {
 	source := sourceImage(t)
