+++ v1/mutate/testdata/whiteout/bar.txt
