+++ v1/mutate/testdata/whiteout/foo.txt
