+++ v1/tarball/image.go
@@ -23,8 +23,6 @@ import (
 	"io"
 	"io/ioutil"
 	"os"
-	"path/filepath"
-	"strings"
 	"sync"
 
 	"github.com/google/go-containerregistry/name"
@@ -33,8 +31,6 @@ import (
 	"github.com/google/go-containerregistry/v1/types"
 	"github.com/google/go-containerregistry/v1/v1util"
 )
-
-const whiteoutPrefix = ".wh."
 
 type image struct {
 	opener        Opener
@@ -179,93 +175,6 @@ func (i *image) loadTarDescriptorAndConfig() error {
 
 func (i *image) RawConfigFile() ([]byte, error) {
 	return i.config, nil
-}
-
-// Flatten takes an image and flattens its layers into a single tarball at the specified path
-// Adapted from https://github.com/google/containerregistry/blob/master/client/v2_2/docker_image_.py#L731
-func Flatten(img v1.Image, tarFilePath string) error {
-	tarFile, err := os.Create(tarFilePath)
-	if err != nil {
-		return fmt.Errorf("Error creating tar file: %v", err)
-	}
-	defer tarFile.Close()
-	tarWriter := tar.NewWriter(tarFile)
-	defer tarWriter.Close()
-
-	fileMap := map[string]bool{}
-
-	layers, err := img.Layers()
-	if err != nil {
-		return fmt.Errorf("Error retrieving image layers: %v", err)
-	}
-	// we iterate through the layers in reverse order because it makes handling
-	// whiteout layers more efficient, since we can just keep track of the removed
-	// files as we see .wh. layers and ignore those in previous layers.
-	for i := len(layers) - 1; i >= 0; i-- {
-		layer := layers[i]
-		layerReader, err := layer.Uncompressed()
-		if err != nil {
-			return fmt.Errorf("Error reading layer contents: %v", err)
-		}
-		tarReader := tar.NewReader(layerReader)
-		for {
-			header, err := tarReader.Next()
-			if err == io.EOF {
-				break
-			}
-			if err != nil {
-				return fmt.Errorf("Error reading tar: %v", err)
-			}
-
-			basename := filepath.Base(header.Name)
-			dirname := filepath.Dir(header.Name)
-			tombstone := strings.HasPrefix(basename, whiteoutPrefix)
-			if tombstone {
-				basename = basename[len(whiteoutPrefix):]
-			}
-
-			// check if we have seen value before
-			name := filepath.Join(dirname, basename)
-			if _, ok := fileMap[name]; ok {
-				continue
-			}
-
-			// check for a whited out parent directory
-			if InWhiteoutDir(fileMap, name) {
-				continue
-			}
-
-			// mark file as handled. non-directory implicitly tombstones
-			// any entries with a matching (or child) name
-			fileMap[name] = tombstone || !(header.Typeflag == tar.TypeDir)
-			if !tombstone {
-				tarWriter.WriteHeader(header)
-				if header.Size > 0 {
-					if _, err := io.Copy(tarWriter, tarReader); err != nil {
-						return err
-					}
-				}
-			}
-		}
-	}
-	return nil
-}
-
-func InWhiteoutDir(fileMap map[string]bool, file string) bool {
-	for {
-		if file == "" {
-			break
-		}
-		dirname := filepath.Dir(file)
-		if file == dirname {
-			break
-		}
-		if val, ok := fileMap[dirname]; ok && val {
-			return true
-		}
-		file = dirname
-	}
-	return false
 }
 
 // tarFile represents a single file inside a tar. Closing it closes the tar itself.
