+++ v1/tarball/image_test.go
@@ -15,100 +15,10 @@
 package tarball
 
 import (
-	"archive/tar"
-	"bytes"
-	"io"
-	"os"
-	"path/filepath"
-	"strings"
 	"testing"
 
 	"github.com/google/go-containerregistry/name"
 )
-
-func TestFlattenWhiteout(t *testing.T) {
-	img, err := ImageFromPath("whiteout_image.tar", nil)
-	if err != nil {
-		t.Errorf("Error loading image: %v", err)
-	}
-	tarPath, _ := filepath.Abs("img.tar")
-	defer os.Remove(tarPath)
-	if err := Flatten(img, tarPath); err != nil {
-		t.Errorf("Error when flattening image: %v", err)
-	}
-	f, err := os.Open(tarPath)
-	if err != nil {
-		t.Errorf("Error when opening tar file for reading: %v", err)
-	}
-	tr := tar.NewReader(f)
-	for {
-		header, err := tr.Next()
-		if err == io.EOF {
-			break
-		}
-		name := header.Name
-		for _, part := range filepath.SplitList(name) {
-			if part == "foo" {
-				t.Errorf("whiteout file found in tar: %v", name)
-			}
-		}
-	}
-}
-
-func TestFlattenOverwrittenFile(t *testing.T) {
-	img, err := ImageFromPath("overwritten_file.tar", nil)
-	if err != nil {
-		t.Fatalf("Error loading image: %v", err)
-	}
-	tarPath, _ := filepath.Abs("img.tar")
-	defer os.Remove(tarPath)
-	if err := Flatten(img, tarPath); err != nil {
-		t.Errorf("Error when flattening image: %v", err)
-	}
-	f, err := os.Open(tarPath)
-	if err != nil {
-		t.Errorf("Error when opening tar file for reading: %v", err)
-	}
-	tr := tar.NewReader(f)
-	for {
-		header, err := tr.Next()
-		if err == io.EOF {
-			break
-		}
-		name := header.Name
-		if strings.Contains(name, "foo.txt") {
-			buf := new(bytes.Buffer)
-			buf.ReadFrom(tr)
-			if strings.Contains(buf.String(), "foo") {
-				t.Errorf("Contents of file were not correctly overwritten")
-			}
-		}
-	}
-}
-
-func TestWhiteoutDir(t *testing.T) {
-	fsMap := map[string]bool{
-		"baz":      true,
-		"red/blue": true,
-	}
-	var tests = []struct {
-		path     string
-		whiteout bool
-	}{
-		{"usr/bin", false},
-		{"baz/foo.txt", true},
-		{"baz/bar/foo.txt", true},
-		{"red/green", false},
-		{"red/yellow.txt", false},
-	}
-
-	for _, tt := range tests {
-		whiteout := InWhiteoutDir(fsMap, tt.path)
-		if whiteout != tt.whiteout {
-			t.Errorf("Whiteout %s: expected %v, but got %v", tt.path, tt.whiteout, whiteout)
-		}
-	}
-}
 
 func TestManifestAndConfig(t *testing.T) {
 	img, err := ImageFromPath("test_image_1.tar", nil)
